refactor: reorganize and add a timeout
