+++ README.md
@@ -1,5 +1,5 @@
 # shoal
 
-Lazy, Python-first, alternative to make, just, doit, and invoke
+Python-first task runner
 
 Documentation can be found on [Github (./docs)](./docs), [PyPi](https://pypi.org/project/shoal/), or [Hosted](https://shoal.kyleking.me/)!
